Leverage the PostgisAuthorityFactory connection bundled in GridGeometryTable, which also allow us to leverage its cache. Changed some logging levels from SEVERE to WARNING or FINE since SEVERE is usually for errors that may prevent the application to work (in this case a particular request may fail, but not the whole application).

+++ postgrid/src/test/java/org/geotools/referencing/factory/wkt/PostgisAuthorityFactoryTest.java
@@ -47,11 +47,10 @@ public class PostgisAuthorityFactoryTest extends DatabaseTest {
     public void testConnected() throws Exception {
         final PostgisAuthorityFactory factory = new PostgisAuthorityFactory(null, database.getConnection());
         assertEquals(Citations.EPSG, factory.getAuthority());
-
-        assertEquals(4326, factory.getPrimaryKey("4326").intValue());
-        assertEquals(4326, factory.getPrimaryKey("EPSG:4326").intValue());
+        assertEquals(4326, factory.getPrimaryKey("4326"));
+        assertEquals(4326, factory.getPrimaryKey("EPSG:4326"));
         try {
-            assertEquals(4326, factory.getPrimaryKey("DUMMY:4326").intValue());
+            assertEquals(4326, factory.getPrimaryKey("DUMMY:4326"));
             fail("Should not find a non-existing authority.");
         } catch (NoSuchAuthorityCodeException e) {
             // This is the expected exception.
